fixed the issue with gas fee data type, uint24 instead uint256

+++ test/BaseTrade.t.sol
@@ -16,7 +16,7 @@ contract BaseTradeTest is Test {
     address quoter = 0x61fFE014bA17989E743c5F6cB21bF9697530B21e;
     address router = 0x68b3465833fb72A70ecDF485E0e4C7bD8665Fc45;
     address pool_factory = 0x1F98431c8aD98523631AE4a59f267346ea31F984 ;
-    uint256 POOL_FEE = 3000;
+    uint24 POOL_FEE = 3000;
     address USDT = 0xdAC17F958D2ee523a2206206994597C13D831ec7;
     address USDC = 0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48;
     address WETH = 0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2;
